srt_announce: Add an option to suppress git send-email auto-cc
Sometimes it's useful to be able to do a 'real' dry run i.e. actually
sending out -rc patches and cover letter to a real e-mail address and
nowhere else.

Doing this with the default auto-cc defeats the purpose though, so add
an option to allow disabling it from the command line.

Example usage:

 srt -d announce -s v3.18.138-rt115 v3.18.138-rt116-rc1

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>
Signed-off-by: Daniel Wagner <wagi@monom.org>