srt_announce: Allow rc releases to also reference release artifacts
The code to generate the rc cover letter assumes there won't be any
references to release tarballs, which isn't the case for everyone, and
so doesn't pass in a lot of the format strings to format().

To change this, create a common map and have the cover-generating code
use that, after appending its own specific strings.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>
Signed-off-by: Daniel Wagner <wagi@monom.org>