srt_upload: Upload incremental update patch to prj_dir/incr, not prj_dir/older
The incremental patch should be uploaded to prj_dir/incr - currently
it's being uploaded to prj_dir/older, which also results in a broken
link in the generated cover letter.

This fix assumes the incremental file name is at position 2 in the
file list, which the other code in srt_upload also assumes.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>
Signed-off-by: Daniel Wagner <wagi@monom.org>