srt_announce: Find correct starting ref for report
We can't just use the old tag as starting point, for a couple of reason.
First we only want to report the changes we did, that means no stable
change log. Second we have a new workflow where the tags changes are not
as previously so we just try to figure out in find_starting_ref() where
our real changes start

Signed-off-by: Daniel Wagner <wagi@monom.org>